FIXED-LOOKS: Download button now goes into a disabled state after an eror response code is received from a putVismodel or postVismodel request.
ENHANCED-LOOKS: Specialised error messages now show up in an alert box in
the dsl panel.

ENHANCED-LOOKS: Reduced the size of the dsl text in the dsl panel.

ENHANCED: raptideAPIHTTP can now be queried for the last response code that was received
for a given request type (i.e., getDSL, postVismodel, putVismodel).

ENHANCED: Events are now fired when an error  response is received for a putVismodel
and postVismodel request.